Removing Logic from FilterVo to GenericUtility and Minor Fix
